+++ src/com/kaylerrenslow/armaplugin/lang/sqf/SQFLocalizeCompletionProvider.java
@@ -1,9 +1,9 @@
 package com.kaylerrenslow.armaplugin.lang.sqf;
 
-import com.intellij.codeInsight.completion.CompletionParameters;
-import com.intellij.codeInsight.completion.CompletionProvider;
-import com.intellij.codeInsight.completion.CompletionResultSet;
+import com.intellij.codeInsight.completion.*;
+import com.intellij.codeInsight.lookup.LookupElement;
 import com.intellij.codeInsight.lookup.LookupElementBuilder;
+import com.intellij.openapi.command.WriteCommandAction;
 import com.intellij.openapi.project.Project;
 import com.intellij.psi.xml.XmlFile;
 import com.intellij.util.ProcessingContext;
@@ -27,15 +27,35 @@ public class SQFLocalizeCompletionProvider extends CompletionProvider<Completion
 	protected void addCompletions(@NotNull CompletionParameters parameters, ProcessingContext context, @NotNull CompletionResultSet result) {
 		Project p = parameters.getOriginalFile().getProject();
 		XmlFile stringTableXml = ArmaPluginUserData.getInstance().getStringTableXml(parameters.getOriginalFile());
-		if (stringTableXml != null) {
-			DomFileElement<StringTableProject> domFileElement = DomManager.getDomManager(p).getFileElement(stringTableXml, StringTableProject.class);
-			if (domFileElement != null) {
-				StringTableProject stringTableProject = domFileElement.getRootElement();
-				List<StringTableKey> allKeys = stringTableProject.getAllKeys();
-				for (StringTableKey key : allKeys) {
-					result.addElement(LookupElementBuilder.create(key.getKey(), key.getID()).appendTailText(" " + key.getContainerPath(), true));
-				}
-			}
+		if (stringTableXml == null) {
+			return;
 		}
+		DomFileElement<StringTableProject> domFileElement = DomManager.getDomManager(p).getFileElement(stringTableXml, StringTableProject.class);
+		if (domFileElement == null) {
+			return;
+		}
+		StringTableProject stringTableProject = domFileElement.getRootElement();
+		List<StringTableKey> allKeys = stringTableProject.getAllKeys();
+		for (StringTableKey key : allKeys) {
+			LookupElementBuilder ele = LookupElementBuilder.create(key, key.getID()).appendTailText(" " + key.getContainerPath(), true);
+			ele = ele.withInsertHandler(new InsertHandler<>() {
+				@Override
+				public void handleInsert(InsertionContext context, LookupElement item) {
+					WriteCommandAction.runWriteCommandAction(
+							context.getProject(), () -> {
+								context.getDocument().replaceString(
+										context.getStartOffset(),
+										context.getTailOffset(),
+										"\"" + item.getLookupString() + "\""
+								);
+							}
+					);
+
+				}
+			});
+			result.addElement(ele);
+		}
+
+
 	}
 }